pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scuba-overlay"
version = "0.1.1"
description = "Generate chroma key-ready dive computer overlay videos from dive logs"
readme = "Readme.md"
requires-python = ">=3.13"
license = {text = "GPL-3.0-or-later"}
authors = [
    {name = "Vladimir"}
]
keywords = ["scuba", "diving", "video", "overlay", "chroma-key", "dive-computer"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Video",
]

dependencies = [
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = []
build = ["pyinstaller>=6.0"]

[project.scripts]
scuba-overlay = "src.main:main"

[tool.setuptools]
packages = []
py-modules = []

[tool.setuptools.package-dir]
"" = "."
